Update gruvbox-dark.sh
Now gruvbox-dark is based on actual Gruvbox, as opposed to the variant of Gruvbox that happened to look good on the original user's machine. IMO, all configuration should be based on standard color palette order (such as "color 0" should be the background color and "color 1" should be semantically "light" red). The previous color scheme used the old palette, but ordered it somewhat randomly, with a blue in the "green" spot, purple in the "blue" spot, etc. Not sure exactly what was happening, but I believe this solves it.