Make Nord Light More Nord
Update Nord Light theme to improve usability and make the black and
grey colors a dark blue.

Regarding usability: Prior to this change, it was extremely difficult
to read green text with Nord Light because the green was such a light
color that the text was hard to discern against the background.

Green text is common in the output of commands like `git diff`, and
many zsh themes use green text to indicate the existence of a command.

This commit uses a darker green that's still in line with the theme.

Regarding update to black and grey: using dark blue and blue-grey
colors for Black and Grey respectively is more in line with the style
choices
made by the Nord theme.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>