Increase contrast of cyan
Cyan is commonly used to display moved sections of text when viewing a
git diff on the command line: the moved-from section is magenta,
and the
moved-to section is cyan. This commit increases the contrast of Cyan
in the Nord Light theme to better cover these use cases, while still
maintaining it's distinguishability as Cyan.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>